+++ src/main/kotlin/com/github/jodiew/sortlines/Util.kt
@@ -45,15 +45,24 @@ data class SortInfo(
     val key: Int? = null,
 ) {
     /** Returns true if the [lines] are sorted according to the [SortInfo], otherwise false */
-    fun isSorted(lines: List<String>): Boolean = if (lines.size <= 1) { true } else {
-        lines.zipWithNext { a, b -> comp(selector(a), selector(b)) }.all { it }
+    fun isSorted(lines: List<String>): Boolean = try { if (lines.size <= 1) { true } else {
+        lines.zipWithNext { a, b ->
+            comp(
+                selector(a),
+                selector(b)
+            )
+        }.all { it }
+    } } catch (e: SortOrderException) {
+        throw e
     }
 
     /** Returns the [lines] in the order specified by the [SortInfo] or null if the [SortInfo.order] is null */
-    fun sorted(lines: List<String>): List<String>? = when (order) {
+    fun sorted(lines: List<String>): List<String>? = try { when (order) {
         SortOrder.ASC -> lines.sortedBy(selector)
         SortOrder.DESC -> lines.sortedByDescending(selector)
         else -> null
+    } } catch (_: SortOrderException) {
+        null
     }
 
     private val type: SortType
@@ -71,16 +80,29 @@ data class SortInfo(
         else -> { _, _ -> true }
     }
 
-    private val selector: (String) -> String = when (type) {
+    private val selector: (String) -> String = try { when (type) {
         SortType.ORDER -> { a -> a }
-        SortType.GROUP -> { a -> if (group != null) { a.getGroup(group) } else { a } }
-        SortType.SPLIT -> { a -> if (split != null && key != null) { a.getSplit(split, key) } else { a } }
+        SortType.GROUP -> { a -> a.getGroup(group!!) }
+        SortType.SPLIT -> { a -> a.getSplit(split!!, key!!) }
+    } } catch (e: SortOrderException) {
+        throw e
     }
 
-    private fun String.getGroup(group: Regex): String = group.find(this)?.groupValues?.getOrNull(1) ?: error("Group pattern not found")
+    private fun String.getGroup(group: Regex): String =
+        group.find(this)?.groupValues?.getOrNull(1)
+            ?: throw SortOrderException(SortBundle.message("inspection.com.github.jodiew.sortlines.error.group"), this)
 
-    private fun String.getSplit(splitPattern: Regex, key: Int): String = trim().split(splitPattern).getOrNull(key) ?: error("Split pattern and key combo not found")
+    private fun String.getSplit(splitPattern: Regex, key: Int): String {
+        val splitGroups = trim().split(splitPattern)
+        if (splitGroups.size <= 1) {
+            throw SortOrderException(SortBundle.message("inspection.com.github.jodiew.sortlines.error.split"), this)
+        } else {
+            return splitGroups.getOrNull(key) ?: throw SortOrderException(SortBundle.message("inspection.com.github.jodiew.sortlines.error.key"), this)
+        }
+    }
 }
+
+class SortOrderException(message: String, val line: String): IllegalStateException(message)
 
 /**
  * Returns the offset of end of the line before an indent change in [text] after [startOffset].
@@ -95,3 +117,11 @@ fun findIndentChangeOffset(text: String, startOffset: Int, initialEndOffset: Int
             match.range.first < initialEndOffset && match.value != initialMatch.value
         }?.range?.first?.minus(1) ?: initialEndOffset
 }
+
+/**
+ * Returns the offset of the start of the next line in the [text] after [startOffset] that has non-whitespace characters.
+ * If that can't be found the [startOffset] is returned
+ */
+fun findLastNewline(text: String, startOffset: Int): Int =
+    Regex("\n(?=[^\n]*\\S)", RegexOption.MULTILINE)
+        .find(text, startOffset)?.range?.first?.inc() ?: startOffset