+++ src/test/testData/quickFix/BlankLinesInSort_before.kt
@@ -1,0 +1,17 @@
+package test
+
+class BlankLinesInSort {
+    // sort: asc
+    val b0 = "b"
+    val a0 = "a"
+
+    val c1 = "c"
+    val e1 = "e"
+    val d1 = "d"
+
+    val xn = "x"
+    val yn = "y"
+    val zn = "z"
+    // sort: end
+    val noSort = "stay"
+}