+++ CHANGELOG.md
@@ -4,6 +4,21 @@
 
 ## [Unreleased]
 ### Added
+- Highlighting for lines where the group or split and key regex can't be found
+- Error message and log message for actions that can't complete a sort
+- "Go to" action in error notification that moves the cursor to the problem sort
+- Improved behaviour for blocks of lines containing blank lines with a `sort: end`
+
+### Fixed
+- Sort not continuing until the `sort: end`, when provided
+
+### Changed
+- Prefix inspection errors with "Sort lines:"
+- Moved usage documentation to a GitHub Pages website
+- Upgrade dependencies: org.jetbrains.intellij.platform 2.10.4, jetbrains/qodana-jvm-community 2025.2, gradle 9.2.1
+
+## [1.0.1] - 2025-10-29
+### Added
 - Media images and usage section in the readme on GitHub
 - Automated generation for lexer and parser
 - Error highlighting for group and split regex
@@ -25,3 +40,7 @@
 - Highlighting for unsorted lines and `Sort lines` quick fix.
 - "Sort" language parsing, highlighting, and simple order completion.
 - Initial scaffold created from [IntelliJ Platform Plugin Template](https://github.com/JetBrains/intellij-platform-plugin-template)
+
+[unreleased]: https://github.com/lollotec/sort-lines/compare/1.0.1...dev
+[1.0.1]: https://github.com/lollotec/sort-lines/releases/tag/1.0.1
+[1.0.0]: https://github.com/lollotec/sort-lines/releases/tag/1.0.0